requirements.txt
absl-py==0.11.0
aspy.yaml==1.3.0
astor==0.8.1
astroid==2.3.3
astunparse==1.6.3
attrs==19.3.0
autopep8==1.5.4
Babel==2.10.1
backcall==0.1.0
bce-python-sdk==0.8.64
bleach==3.1.4
bottle==0.12.18
cached-property==1.5.2
cachetools==4.2.1
certifi==2020.6.20
cffi==1.15.0
chardet==3.0.4
clang==5.0
click==8.1.2
colorama==0.4.3
cycler==0.10.0
Cython==0.29.28
decorator==4.4.2
defusedxml==0.6.0
distlib==0.3.4
easydict==1.9
einops==0.3.0
entrypoints==0.3
et-xmlfile==1.0.1
fanstatic==1.1
filelock==3.6.0
flake8==4.0.1
Flask==2.1.1
Flask-Babel==2.0.0
flatbuffers==1.12
future==0.18.2
gast==0.4.0
geojson==2.5.0
gevent==21.12.0
google-auth==1.25.0
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
greenlet==1.1.2
grpcio==1.45.0
h5py==3.1.0
hmmlearn==0.2.5
idna==2.10
imageio==2.9.0
importlib-metadata==4.11.3
ipykernel==5.2.0
ipython==7.13.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
isort==4.3.21
itsdangerous==2.1.2
jdcal==1.4.1
jedi==0.16.0
jieba==0.42.1
Jinja2==3.1.1
joblib==1.0.0
js==1.0
Js2Py==0.70
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.11
jupyter-console==6.2.0
jupyter-core==4.7.0
keras==2.8.0
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
kiwisolver==1.2.0
lap==0.4.0
lazy-object-proxy==1.4.3
llvmlite==0.36.0
Markdown==3.3.3
MarkupSafe==2.1.1
marshmallow==3.7.0
matplotlib==3.3.0
mccabe==0.6.1
mistune==0.8.4
motmetrics==1.2.5
mpmath==1.1.0
munch==2.5.0
natsort==8.1.0
nbconvert==5.6.1
nbformat==5.0.5
networkx==2.5.1
nltk==3.5
nodeenv==1.6.0
notebook==6.0.3
numba==0.53.1
numpy==1.19.3
oauthlib==3.1.0
opencv-contrib-python==4.3.0.38
opencv-python==4.4.0.46
openni==2.3.0
openpyxl==3.0.5
opt-einsum==3.3.0
paddlepaddle-gpu==2.2.2.post112
paddleslim==2.2.2
pandas==1.2.0
pandocfilters==1.4.2
parso==0.6.2
patsy==0.5.1
pickleshare==0.7.5
Pillow==7.2.0
platformdirs==2.5.2
platformio==4.3.4
pre-commit==0.14.3
prometheus-client==0.7.1
prompt-toolkit==3.0.5
protobuf==3.14.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybind11==2.5.0
pycocotools==2.0.4
pycodestyle==2.8.0
pycparser==2.21
pycryptodome==3.14.1
pydot==1.4.2
pyelftools==0.26
pyflakes==2.4.0
pygame==1.9.6
Pygments==2.6.1
pyjsparser==2.7.1
pylint==2.4.4
pymssql==2.2.5
pyparsing==2.4.7
PyQt5==5.15.2
PyQt5-sip==12.8.1
PyQt5-stubs==5.14.2.2
pyrsistent==0.16.0
pyserial==3.4
python-dateutil==2.8.1
pytz==2020.5
PyWavelets==1.1.1
pywin32==227
pywinpty==0.5.7
PyYAML==5.4.1
pyzmq==19.0.0
qtconsole==4.7.2
rawpy==0.16.0
regex==2020.11.13
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.7
scikit-image==0.18.1
scikit-learn==0.23.2
scipy==1.5.4
seaborn==0.11.1
semantic-version==2.8.5
Send2Trash==1.5.0
Shapely==1.8.1.post1
shellcheck-py==0.8.0.4
shutilwhich==1.1.0
six==1.15.0
sklearn==0.0
statsmodels==0.12.1
sympy==1.7.1
tabulate==0.8.7
taichi==0.6.24
taichi-glsl==0.0.8
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow-estimator==2.8.0
tensorflow-gpu==2.6.0
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
tf-slim==1.1.0
threadpoolctl==2.1.0
tifffile==2021.6.6
toml==0.10.1
torch==1.8.0
torchvision==0.9.0
tornado==6.0.4
tqdm==4.56.0
traitlets==4.3.3
typing-extensions==3.7.4.3
tzlocal==2.1
urllib3==1.25.9
utils==1.0.1
virtualenv==20.14.1
visualdl==2.2.3
wcwidth==0.1.9
webencodings==0.5.1
WebOb==1.8.6
Werkzeug==2.1.1
widgetsnbextension==3.5.1
wordcloud==1.8.1
wrapt==1.12.1
xmltodict==0.12.0
zipp==3.8.0
zope.event==4.5.0
zope.interface==5.4.0

paddlers~=0.0.1